Add initial SPIs and providers
This forms a stable starting point for #1, #2, and #3 SPIs along with providers for #8, #9, and #10. These consist of an AEAD cipher SPI, a key exchange SPI, and a signature SPI, along with providers based on BouncyCastle, Java 11, and Java 15 respectively. A small number of smoke tests are included from the relevant RFCs.